requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confusable-homoglyphs==3.2.0
distlib==0.3.3
Django==3.2.8
django-admin-sortable2==1.0.2
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-jinja==2.9.1
django-jinja-bootstrap-form==4.3.4
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
django-redis==5.0.0
django-registration==3.2
django-rosetta==0.9.7
django-spurl==0.6.7
django-starfield==1.0.post2
django-treebeard==4.5.1
djangorestframework==3.12.4
filelock==3.3.1
idna==3.3
Jinja2==3.0.2
kombu==5.1.0
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==2.0.1
packaging==21.0
phonenumbers==8.12.35
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
polib==1.1.1
prompt-toolkit==3.0.20
py==1.10.0
pyparsing==2.4.7
python-dotenv==0.19.1
python-http-client==3.3.3
pytz==2021.3
redis==3.5.3
requests==2.26.0
sendgrid==3.6.5
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.0
toml==0.10.2
tox==3.24.4
urllib3==1.26.7
URLObject==2.4.3
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
